backend/requirements.txt
gunicorn>=21.0.0
uvicorn>=0.30.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.77.0
anyio==4.12.1
attrs==25.4.0
bcrypt==3.2.2
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
ecdsa==0.19.1
email-validator==2.3.0
environs==14.5.0
fastapi==0.128.0
fastapi-cors==0.0.6
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==4.2.1
multidict==6.7.1
narwhals==2.16.0
numpy==2.4.2
openai==2.16.0
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pillow==12.1.0
plotly==6.5.2
propcache==0.4.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyarrow==23.0.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
pydeck==0.9.1
pypdf==6.6.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.50.0
streamlit==1.54.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.2
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
yarl==1.22.0
slowapi>=0.1.9
pymysql==1.1.2
alembic==1.18.4


